Move namespaces to FastJetOptions.h
